app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-keycloak==3.7.0
requests==2.31.0
pydantic-settings==2.0.3
python-multipart==0.0.6